refactor: separate form/UI state from API mutation in useCreateLetterhead
Move useState (PDF files, margins), useForm, resetForm, and onSubmit
logic out of the useCreateLetterhead API hook and into the
LetterheadFormContent component where they belong.

The API hook now only handles the mutation concern, returning
{ createLetterhead, isCreating } — consistent with useUpdateLetterhead
and useDeleteLetterhead.

Applied via @cursor push command
